+++ src/screen.rs
@@ -1,12 +1,12 @@
 pub trait Screen
 {
     fn Clear(&mut self);
-    fn PlotPoint(&mut self, x : f32, y :f32);
-    fn PlotLine(&mut self, x1 : f32, y1 :f32, x2 : f32, y2 : f32);
-    fn PlotCircle(&mut self, x : f32, y: f32, r: f32);
-    fn PlotRectangle(&mut self, x1 : f32, y1 :f32, x2 : f32, y2 : f32);
-    fn Position(&mut self, x : f32, y : f32);
-    fn Zoom(&mut self, zoom : f32);
+    fn PlotPoint(&mut self, x : f64, y :f64);
+    fn PlotLine(&mut self, x1 : f64, y1 :f64, x2 : f64, y2 : f64);
+    fn PlotCircle(&mut self, x : f64, y: f64, r: f64);
+    fn PlotRectangle(&mut self, x1 : f64, y1 :f64, x2 : f64, y2 : f64);
+    fn Position(&mut self, x : f64, y : f64);
+    fn Zoom(&mut self, zoom : f64);
     fn Draw(&mut self);
     fn set_palette(&mut self, palette : i32);
     fn TextOutputter(&mut self, output : Box<dyn TextOutputter>);
